Extract task_contract_recovery.rs from turn.rs (parent #680)
Move the TaskContract recovery action / target planners out of
`turn.rs` into a focused sibling module so `turn.rs` keeps shrinking
toward responsibility-focused units (parent meta-issue #680).

Two production helpers + one private adapter extracted as free
functions taking `&mut Agent` / `&Agent` (matching `actor_loop_flow` /
`reply_retry` / earlier vertical-slice precedent):

- `task_contract_recovery_action` (production chokepoint: routes
  between `RunVerifier` / `Continue` / `Done` per artifact state +
  repair state + completion probe; uses the `artifact_state_projection`
  SSOT for ledger / legacy alignment; calls `project_probe::probe_completion`
  to short-circuit speculative `Continue` decisions)
- `task_contract_recovery_target` (`Continue { missing }` →
  `RecoveryTargetHint` for the first missing role; tries scaffold
  candidate → scope-internal Owned workspace artifact → synthesised
  conventional implementation path)
- `task_contract_repair_state` (private adapter to
  `super::repair_job::task_contract_repair_state_from_job`)

Behavior unchanged: same verifier-repair readiness gate, same
artifact-state-projection consumption, same completion-probe
short-circuit + masked `agent.completion_probe.decision` emit shape,
same target-hint priority chain (scaffold → existing Owned →
synthesised) with `NestedTestAdmission::default()` conservative guard
preserved.

`pub(super)` limited / no facade re-export (DR3-001). Callers updated
in `actor_loop_flow.rs` (3 sites), `set_artifact_recovery_target.rs`
(1 site), and `progress_tests.rs` (1 test site) to the free-fn form.

LOC delta:
- turn.rs: 1,646 → 1,471 (-175)
- new module: +204

3,114 lib tests pass; `cargo fmt` / `cargo clippy --lib --all-targets
-- -D warnings` clean.
